change multi parameter method for mapper

+++ src/main/java/com/loha/flippedclassroom/controller/StudentController.java
@@ -36,7 +36,7 @@ public class StudentController {
 
     @PostMapping(value = "/activation")
     @ResponseBody
-    public ResponseEntity activateStudent(@ModelAttribute("curStudentId") Integer studentId, String password,String email) throws Exception{
+    public ResponseEntity activateStudent(@ModelAttribute("curStudentId") long studentId, String password,String email) throws Exception{
         studentService.activateStudent(password,email,studentId);
         return new ResponseEntity(HttpStatus.ACCEPTED);
     }
@@ -53,13 +53,13 @@ public class StudentController {
     }
 
     @GetMapping(value = "/course")
-    public String getMyCourse(@ModelAttribute("curStudentId") Integer studentId,Model model) throws Exception{
+    public String getMyCourse(@ModelAttribute("curStudentId") long studentId,Model model) throws Exception{
         model.addAttribute("courseAndKlassList",studentService.getCourseAndKlass(studentId));
         return "student/myCourse";
     }
 
     @GetMapping(value = "/setting")
-    public String getSetting(@ModelAttribute("curStudentId") Integer studentId,Model model)throws Exception{
+    public String getSetting(@ModelAttribute("curStudentId") long studentId,Model model)throws Exception{
         model.addAttribute("student",studentService.getStudentById(studentId));
         return "student/settings";
     }
@@ -71,7 +71,7 @@ public class StudentController {
 
     @PostMapping(value = "/setting/modifyEmail")
     @ResponseBody
-    public ResponseEntity modifyEmail(@ModelAttribute("curStudentId") Integer studentId,String email) throws Exception{
+    public ResponseEntity modifyEmail(@ModelAttribute("curStudentId") long studentId,String email) throws Exception{
         studentService.modifyStudentEmail(studentId,email);
         return new ResponseEntity(HttpStatus.ACCEPTED);
     }
@@ -83,19 +83,19 @@ public class StudentController {
 
     @PostMapping(value = "/setting/modifyPwd")
     @ResponseBody
-    public ResponseEntity modifyPwd(@ModelAttribute("curStudentId") Integer studentId,String password) throws Exception{
+    public ResponseEntity modifyPwd(@ModelAttribute("curStudentId") long studentId,String password) throws Exception{
         studentService.modifyStudentPwdById(studentId,password);
         return new ResponseEntity(HttpStatus.ACCEPTED);
     }
 
     @PostMapping(value = "/course/info")
-    public String courseInfo(Integer courseId,Model model) throws Exception{
+    public String courseInfo(long courseId,Model model) throws Exception{
         model.addAttribute("course",studentService.getCourseById(courseId));
         return "student/courseInfo";
     }
 
     @PostMapping(value = "/course/score")
-    public String getMyScoreInfo(@ModelAttribute("curStudentId") Integer studentId,Integer klassId,Integer courseId,Model model) throws Exception{
+    public String getMyScoreInfo(@ModelAttribute("curStudentId") long studentId,long klassId,long courseId,Model model) throws Exception{
         model.addAttribute("scoreList",studentService.getMyScore(klassId,courseId,studentId));
         return "student/scoreInfo";
     }
